[license] Use license sets and correct license header usage
Was having issues with this for long time, was very obscure.  The license.header here was propagated obviously to the internal plugin.  In past that didn't harm anything, just duplicate in setup, now it causes a problem as it hits the deprecated code.  At one time, I had older license style still present before removing it entirely for newer style.  So move the header more directly on setup.  If it ever needs to change again in similar way, need to avoid name 'license.header' in this way.